Re-factor TaskFragment in to separate classes per task type
